+++ R/step-mutate.R
@@ -16,7 +16,9 @@ step_mutate <- function(parent, new_vars = list(), nested = FALSE) {
 
 dt_call.dtplyr_step_mutate <- function(x, needs_copy = x$needs_copy) {
   # i is always empty because we never mutate a subset
-  if (!x$nested) {
+  if (is_empty(x$new_vars)) {
+    j <- quote(.SD)
+  } else if (!x$nested) {
     j <- call2(":=", !!!x$new_vars)
   } else {
     mutate_list <- mutate_nested_vars(x$new_vars)
@@ -44,12 +46,18 @@ mutate_nested_vars <- function(mutate_vars) {
 #' Create and modify columns
 #'
 #' This is a method for the dplyr [mutate()] generic. It is translated to
-#' the `j` argument of `[.data.table`, using `:=` to modify "in place".
+#' the `j` argument of `[.data.table`, using `:=` to modify "in place". If 
+#' `.before` or `.after` is provided, the new columns are relocated with a call
+#' to [data.table::setcolorder()].
 #'
 #' @param .data A [lazy_dt()].
 #' @param ... <[data-masking][dplyr::dplyr_data_masking]> Name-value pairs.
 #'   The name gives the name of the column in the output, and the value should
 #'   evaluate to a vector.
+#' @param .before,.after \Sexpr[results=rd]{lifecycle::badge("experimental")}
+#'   <[`tidy-select`][dplyr_tidy_select]> Optionally, control where new columns
+#'   should appear (the default is to add to the right hand side). See
+#'   [relocate()] for more details.
 #' @importFrom dplyr mutate
 #' @export
 #' @examples
@@ -63,14 +71,25 @@ mutate_nested_vars <- function(mutate_vars) {
 #' # are used in the same expression
 #' dt %>%
 #'   mutate(x1 = x + 1, x2 = x1 + 1)
-mutate.dtplyr_step <- function(.data, ...) {
+mutate.dtplyr_step <- function(.data, ...,
+                               .before = NULL, .after = NULL) {
   dots <- capture_dots(.data, ...)
   if (is_null(dots)) {
     return(.data)
   }
 
   nested <- nested_vars(.data, dots, .data$vars)
-  step_mutate(.data, dots, nested)
+  out <- step_mutate(.data, dots, nested)
+
+  .before <- enquo(.before)
+  .after <- enquo(.after)
+  if (!quo_is_null(.before) || !quo_is_null(.after)) {
+    # Only change the order of new columns
+    new <- setdiff(names(dots), .data$vars)
+    out <- relocate(out, !!new, .before = !!.before, .after = !!.after)
+  }
+
+  out
 }
 
 #' @export