+++ R/tidyeval-across.R
@@ -8,7 +8,7 @@ dt_squash_across <- function(call, env, data, j = j) {
 
   tbl <- simulate_vars(data, drop_groups = TRUE)
   .cols <- call$.cols %||% expr(everything())
-  locs <- tidyselect::eval_select(.cols, tbl, allow_rename = FALSE)
+  locs <- tidyselect::eval_select(.cols, tbl)
   cols <- syms(names(tbl))[locs]
 
   funs <- across_funs(call$.fns, env, data, j = j)
@@ -27,9 +27,7 @@ dt_squash_across <- function(call, env, data, j = j) {
   }
 
   .names <- eval(call$.names, env)
-  if (!is.null(call$.fns)) {
-    names(out) <- across_names(names(tbl)[locs], names(funs), .names, env)
-  }
+  names(out) <- across_names(names(locs), names(funs), .names, env)
   out
 }
 
@@ -43,7 +41,8 @@ dt_squash_if <- function(call, env, data, j = j, reduce = "&") {
   call <- match.call(dplyr::if_any, call, expand.dots = FALSE, envir = env)
 
   tbl <- simulate_vars(data, drop_groups = TRUE)
-  locs <- tidyselect::eval_select(call$.cols, tbl, allow_rename = FALSE)
+  .cols <- call$.cols %||% expr(everything())
+  locs <- tidyselect::eval_select(.cols, tbl, allow_rename = FALSE)
   cols <- syms(names(tbl))[locs]
 
   fun <- across_fun(call$.fns, env, data, j = j)
@@ -60,8 +59,8 @@ dt_squash_if <- function(call, env, data, j = j, reduce = "&") {
 across_funs <- function(funs, env, data, j = TRUE) {
   if (is.null(funs)) {
     list(function(x, ...) x)
-  } else if (is_symbol(funs)) {
-    set_names(list(across_fun(funs, env, data, j = j)), as.character(funs))
+  } else if (is_symbol(funs) || is_function(funs)) {
+    set_names(list(across_fun(funs, env, data, j = j)), as_label(funs))
   } else if (is.character(funs)) {
     names(funs)[names2(funs) == ""] <- funs
     lapply(funs, across_fun, env, data, j = j)
@@ -75,19 +74,20 @@ across_funs <- function(funs, env, data, j = TRUE) {
     funs <- eval(funs, env)
     across_funs(funs, NULL)
   } else {
-    abort("`.fns` argument to dtplyr::across() must be a NULL, a function name, formula, or list")
+    abort("`.fns` argument to dtplyr::across() must be a NULL, a function, formula, or list")
   }
 }
 
 across_fun <- function(fun, env, data, j = TRUE) {
-  if (is_symbol(fun) || is_string(fun)) {
+  if (is_symbol(fun) || is_string(fun) ||
+    is_call(fun, "function") || is_function(fun)) {
     function(x, ...) call2(fun, x, ...)
   } else if (is_call(fun, "~")) {
     call <- dt_squash_formula(fun, env, data, j = j, replace = quote(!!.x))
     function(x, ...) expr_interp(call, child_env(emptyenv(), .x = x))
   } else {
     abort(c(
-      ".fns argument to dtplyr::across() must contain a function name or a formula",
+      ".fns argument to dtplyr::across() must contain a function or a formula",
       x = paste0("Problem with ", expr_deparse(fun))
     ))
   }
@@ -96,20 +96,23 @@ across_fun <- function(fun, env, data, j = TRUE) {
 dt_squash_formula <- function(x, env, data, j = TRUE, replace = quote(!!.x)) {
   call <- f_rhs(x)
   call <- replace_dot(call, replace)
-  call <- dt_squash_call(call, env, data, j = j)
+  if (is_call(call)) {
+    call <- dt_squash_call(call, env, data, j = j)
+  }
   call
 }
 
 across_names <- function(cols, funs, names = NULL, env = parent.frame()) {
-  if (length(funs) == 1) {
+  n_reps <- if (is_empty(funs)) 1 else length(funs)
+  if (n_reps == 1) {
     names <- names %||% "{.col}"
   } else {
     names <- names %||% "{.col}_{.fn}"
   }
 
   glue_env <- child_env(env,
-    .col = rep(cols, each = length(funs)),
-    .fn = rep(funs %||% seq_along(funs), length(cols))
+    .col = rep(cols, each = n_reps),
+    .fn = rep(funs %||% seq_len(n_reps), length(cols))
   )
   glue::glue(names, .envir = glue_env)
 }