+++ R/tidyeval.R
@@ -16,7 +16,7 @@ dt_eval <- function(x) {
 dt_funs <- c(
   "between", "CJ", "copy", "data.table", "dcast", "melt", "nafill",
   "fcase", "fcoalesce", "fifelse", "fintersect", "frank", "frankv", "fsetdiff", "funion",
-  "setcolorder", "setnames", "tstrsplit", "uniqueN"
+  "setcolorder", "setnames", "shift", "tstrsplit", "uniqueN"
 )
 add_dt_wrappers <- function(env) {
   env_bind(env, !!!env_get_list(ns_env("data.table"), dt_funs))
@@ -130,6 +130,9 @@ dt_squash_call <- function(x, env, data, j = TRUE) {
   } else if (is_call(x, "cur_group_rows")) {
     quote(.I)
   } else if (is_call(x, "desc")) {
+      if (!has_length(x, 2L)) {
+        abort("`desc()` expects exactly one argument.")
+      }
     x[[1]] <- sym("-")
     x[[2]] <- get_expr(x[[2]])
     x
@@ -143,6 +146,28 @@ dt_squash_call <- function(x, env, data, j = TRUE) {
     x[[1]] <- quote(fifelse)
     x[-1] <- lapply(x[-1], dt_squash, env, data, j = j)
     x
+  } else if (is_call(x, c("lag", "lead"))) {
+    if (is_call(x, "lag")) {
+      type <- "lag"
+      call <- match.call(dplyr::lag, x)
+    } else {
+      type <- "lead"
+      call <- match.call(dplyr::lead, x)
+    }
+    shift_call <- call2("shift", x[[2]])
+    if (!is_null(call$n)) {
+      shift_call$n <- call$n
+    }
+    if (!is_null(call$default)) {
+      shift_call$fill <- call$default
+    }
+    if (!is_null(call$order_by)) {
+      abort(
+        glue::glue("The `order_by` argument of `{type}()` is not supported by dtplyr")
+      )
+    }
+    shift_call$type <- type
+    shift_call
   } else if (is_call(x, "n", n = 0)) {
     quote(.N)
   } else if (is_call(x, "n_distinct")) {