+++ tests/testthat/test-tidyeval-across.R
@@ -3,7 +3,7 @@ test_that("across() translates NULL", {
 
   expect_equal(
     capture_across(dt, across(a:b)),
-    list(expr(a), expr(b))
+    list(a = expr(a), b = expr(b))
   )
 })
 
@@ -12,11 +12,11 @@ test_that("across() drops groups", {
 
   expect_equal(
     capture_across(group_by(dt, a), across(everything())),
-    list(expr(b))
+    list(b = expr(b))
   )
   expect_equal(
     capture_across(group_by(dt, b), across(everything())),
-    list(expr(a))
+    list(a = expr(a))
   )
 })
 
@@ -58,6 +58,15 @@ test_that("across() translates functions", {
   )
 })
 
+test_that("across() captures anonymous functions", {
+  dt <- lazy_dt(data.frame(a = 1))
+
+  expect_equal(
+   capture_across(dt, across(a, function(x) log(x))),
+   list(a = call2(function(x) log(x), quote(a)))
+  )
+})
+
 test_that("dots are translated too", {
   fun <- function() {
     dt <- lazy_dt(data.frame(a = 1, b = 2))
@@ -83,6 +92,11 @@ test_that("across() translates formulas", {
   )
 
   expect_equal(
+    capture_across(dt, across(a:b, ~2)),
+    exprs(a = 2, b = 2)
+  )
+
+  expect_equal(
     capture_across(dt, across(a:b, list(~log(.x)))),
     exprs(a = log(a), b = log(b))
   )
@@ -94,6 +108,56 @@ test_that("across() gives informative errors", {
     capture_across(dt, across(a, 1))
     capture_across(dt, across(a, list(1)))
   })
+})
+
+test_that("across() can use named selections", {
+  dt <- lazy_dt(data.frame(x = 1, y = 2))
+
+  # no fns
+  expect_equal(
+    capture_across(dt, across(c(a = x, b = y))),
+    list(a = quote(x), b = quote(y))
+  )
+  expect_equal(
+    capture_across(dt, across(all_of(c(a = "x", b = "y")))),
+    list(a = quote(x), b = quote(y))
+  )
+
+  # one fn
+  expect_equal(
+    capture_across(dt, across(c(a = x, b = y), mean)),
+    list(a = quote(mean(x)), b = quote(mean(y)))
+  )
+  expect_equal(
+    capture_across(dt, across(all_of(c(a = "x", b = "y")), mean)),
+    list(a = quote(mean(x)), b = quote(mean(y)))
+  )
+
+  # multiple fns
+  expect_equal(
+    capture_across(dt, across(c(a = x, b = y), list(mean, nm = sum))),
+    list(
+      a_mean = quote(mean(x)), a_nm = quote(sum(x)),
+      b_mean = quote(mean(y)), b_nm = quote(sum(y))
+    )
+  )
+  expect_equal(
+    capture_across(dt, across(all_of(c(a = "x", b = "y")), list(mean, nm = sum))),
+    list(
+      a_mean = quote(mean(x)), a_nm = quote(sum(x)),
+      b_mean = quote(mean(y)), b_nm = quote(sum(y))
+    )
+  )
+
+})
+
+test_that("across() can handle empty selection", {
+  dt <- lazy_dt(data.table(x = 1, y = 2), "DT")
+
+  expect_equal(
+    dt %>% mutate(across(character(), c)) %>% show_query(),
+    expr(copy(DT)[, .SD])
+  )
 })